move nanpercentile to nanops
